warning free build on OS X 10.5.8
git-svn-id: svn://coreboot.org/openbios/trunk/fcode-utils-devel@759 f158a5a8-5612-0410-a976-696ce0be7e32

+++ detok/pcihdr.c
@@ -461,7 +461,7 @@ void handle_pci_filler(u8 * filler_ptr)
 	} else {
 		sprintf(temp_buf, "PCI Image padding-field of %d bytes "
 			"had first non-zero byte at offset %ld",
-			filler_len, scan_ptr - filler_ptr);
+			filler_len, (unsigned long)(scan_ptr - filler_ptr));
 	}
 	printremark(temp_buf);
 }
